Refactor createAuthProfileExp function import path and add new utility file for authentication profile creation
