+++ src/common/utilities/PacAuthUtil.ts
